+++ stapleui/pubspec.yaml
@@ -10,6 +10,7 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  http: ^1.2.2
   provider: ^6.1.2
   web_socket_channel: ^2.1.0
